数据结构/set/src/com/set/TreeSet.java
package com.set;

import com.tree.rbTree.RBTree;

public class TreeSet<E> implements Set<E> {

    private final RBTree<E> tree = new RBTree<E>();

    @Override
    public int size() {
        return tree.size();
    }

    @Override
    public boolean isEmpty() {
        return tree.isEmpty();
    }

    @Override
    public void clear() {
        tree.clear();
    }

    @Override
    public boolean contains(E element) {
        return tree.contains(element);
    }

    @Override
    public void add(E element) {
        tree.add(element);
    }

    @Override
    public void remove(E element) {
        tree.remove(element);
    }

    @Override
    public void traversal(Visitor<E> visitor) {
        if (visitor == null) return;
        tree.levelOrderTraversal();
    }
}
